delete old version of vit
