+++ libai/models/__init__.py
@@ -16,8 +16,7 @@
 from .bert_model import BertForPreTraining, BertModel
 from .build import build_graph, build_model
 from .t5_model import T5ForPreTraining, T5Model
-# from .vision_transformer import VisionTransformer
-from .vit_libai import VisionTransformer
+from .vision_transformer import VisionTransformer
 
 __all__ = [
     "build_model",
