+++ libai/models/vision_transformer.py
@@ -13,355 +13,142 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import math
-from collections import OrderedDict
-from functools import partial
-
 import oneflow as flow
 import oneflow.nn as nn
-from flowvision.layers.blocks import Mlp, PatchEmbed
-from flowvision.layers.weight_init import lecun_normal_, trunc_normal_
-from flowvision.models.helpers import named_apply
+from flowvision.layers.weight_init import trunc_normal_
 
+from libai.layers import (
+    PatchEmbedding, 
+    TransformerLayer, 
+    LayerNorm,
+    Linear,
+)
 from libai.config.config import configurable
+import libai.utils.distributed as dist
 
 from .build import MODEL_ARCH_REGISTRY
-
-
-def drop_path(x, drop_prob: float = 0.5, training: bool = False):
-    if drop_prob == 0.0 or not training:
-        return x
-    keep_prob = 1 - drop_prob
-    shape = (x.shape[0],) + (1,) * (x.ndim - 1)  # work with diff dim tensors, not just 2D ConvNets
-    random_tensor = (
-        flow.rand(*shape, dtype=x.dtype, sbp=flow.sbp.broadcast, placement=x.placement).to_global(
-            sbp=x.sbp
-        )
-        + keep_prob
-    )
-    random_tensor = random_tensor.floor()  # binarize
-    output = x.div(keep_prob) * random_tensor
-    return output
-
-
-class DropPath(nn.Module):
-    """Drop paths (Stochastic Depth) per sample  (when applied in main path of residual blocks)."""
-
-    def __init__(self, drop_prob=None):
-        super(DropPath, self).__init__()
-        self.drop_prob = drop_prob
-
-    def forward(self, x):
-        return drop_path(x, self.drop_prob, self.training)
-
-
-class Attention(nn.Module):
-    def __init__(self, dim, num_heads=8, qkv_bias=False, attn_drop=0.0, proj_drop=0.0):
-        super().__init__()
-        self.num_heads = num_heads
-        head_dim = dim // num_heads
-        self.scale = head_dim ** -0.5
-
-        self.qkv = nn.Linear(dim, dim * 3, bias=qkv_bias)
-        self.attn_drop = nn.Dropout(attn_drop)
-        self.proj = nn.Linear(dim, dim)
-        self.proj_drop = nn.Dropout(proj_drop)
-
-    def forward(self, x):
-        B, N, C = x.shape
-        qkv = (
-            self.qkv(x).reshape(B, N, 3, self.num_heads, C // self.num_heads).permute(2, 0, 3, 1, 4)
-        )
-        # TODO supported tensor.unbind in oneflow
-        # q, k, v = qkv.unbind(0)
-        q, k, v = qkv[0], qkv[1], qkv[2]
-
-        attn = (q @ k.transpose(-2, -1)) * self.scale
-        attn = attn.softmax(dim=-1)
-        attn = self.attn_drop(attn)
-
-        x = (attn @ v).transpose(1, 2).reshape(B, N, C)
-        x = self.proj(x)
-        x = self.proj_drop(x)
-        return x
-
-
-class Block(nn.Module):
-    def __init__(
-        self,
-        dim,
-        num_heads,
-        mlp_ratio=4.0,
-        qkv_bias=False,
-        drop=0.0,
-        attn_drop=0.0,
-        drop_path=0.0,
-        act_layer=nn.GELU,
-        norm_layer=nn.LayerNorm,
-    ):
-        super().__init__()
-        self.norm1 = norm_layer(dim)
-        self.attn = Attention(
-            dim,
-            num_heads=num_heads,
-            qkv_bias=qkv_bias,
-            attn_drop=attn_drop,
-            proj_drop=drop,
-        )
-        # NOTE: drop path for stochastic depth, we shall see if this is better than dropout here
-        self.drop_path = DropPath(drop_path) if drop_path > 0.0 else nn.Identity()
-        self.norm2 = norm_layer(dim)
-        mlp_hidden_dim = int(dim * mlp_ratio)
-        self.mlp = Mlp(
-            in_features=dim,
-            hidden_features=mlp_hidden_dim,
-            act_layer=act_layer,
-            drop=drop,
-        )
-
-    def forward(self, x):
-        x = x + self.drop_path(self.attn(self.norm1(x)))
-        x = x + self.drop_path(self.mlp(self.norm2(x)))
-        return x
 
 
 @MODEL_ARCH_REGISTRY.register()
 class VisionTransformer(nn.Module):
     """Vision Transformer
-    An OneFlow impl of: `An Image is Worth 16x16 Words: Transformers for Image Recognition at Scale`
+    LiBai impl of: `An Image is Worth 16x16 Words: Transformers for Image Recognition at Scale`
         - https://arxiv.org/abs/2010.11929
-    Includes distillation token & head support for `DeiT: Data-efficient Image Transformers`
-        - https://arxiv.org/abs/2012.12877
     """
-
+    
     @configurable
     def __init__(
         self,
         img_size=224,
         patch_size=16,
         in_chans=3,
-        num_classes=1000,
-        embed_dim=768,
+        embed_dim=192,
         depth=12,
-        num_heads=12,
+        num_heads=3,
         mlp_ratio=4.0,
-        qkv_bias=True,
-        representation_size=None,
-        distilled=False,
         drop_rate=0.0,
         attn_drop_rate=0.0,
         drop_path_rate=0.0,
-        embed_layer=PatchEmbed,
-        norm_layer=None,
-        act_layer=None,
-        weight_init="",
-        loss_func=None,
+        num_classes=1000,
+        loss_func = None,
     ):
-        """
-        Args:
-            img_size (int, tuple): input image size
-            patch_size (int, tuple): patch size
-            in_chans (int): number of input channels
-            num_classes (int): number of classes for classification head
-            embed_dim (int): embedding dimension
-            depth (int): depth of transformer
-            num_heads (int): number of attention heads
-            mlp_ratio (int): ratio of mlp hidden dim to embedding dim
-            qkv_bias (bool): enable bias for qkv if True
-            representation_size (Optional[int]): enable and set representation layer (pre-logits)
-                                                 to this value if set
-            distilled (bool): model includes a distillation token and head as in DeiT models
-            drop_rate (float): dropout rate
-            attn_drop_rate (float): attention dropout rate
-            drop_path_rate (float): stochastic depth rate
-            embed_layer (nn.Module): patch embedding layer
-            norm_layer: (nn.Module): normalization layer
-            weight_init: (str): weight init scheme
-        """
         super().__init__()
         self.num_classes = num_classes
-        self.num_features = (
-            self.embed_dim
-        ) = embed_dim  # num_features for consistency with other models
-        self.num_tokens = 2 if distilled else 1
-        norm_layer = norm_layer or partial(nn.LayerNorm, eps=1e-6)
-        act_layer = act_layer or nn.GELU
-
-        self.patch_embed = embed_layer(
+        self.patch_embed = PatchEmbedding(
             img_size=img_size,
             patch_size=patch_size,
             in_chans=in_chans,
             embed_dim=embed_dim,
         )
+        ffn_size = int(embed_dim * mlp_ratio)
         num_patches = self.patch_embed.num_patches
+        self.cls_token = nn.Parameter(flow.zeros(1, 1, embed_dim, sbp=dist.get_nd_sbp([flow.sbp.broadcast, flow.sbp.broadcast]),
+                                            placement=dist.get_layer_placement(0)))
+        self.pos_embed = nn.Parameter(flow.zeros(1, num_patches + 1, embed_dim, sbp=dist.get_nd_sbp([flow.sbp.broadcast, flow.sbp.broadcast]),
+                                                                                placement=dist.get_layer_placement(0)))
 
-        self.cls_token = nn.Parameter(flow.zeros(1, 1, embed_dim))
-        self.dist_token = nn.Parameter(flow.zeros(1, 1, embed_dim)) if distilled else None
-        self.pos_embed = nn.Parameter(flow.zeros(1, num_patches + self.num_tokens, embed_dim))
         self.pos_drop = nn.Dropout(p=drop_rate)
-
         dpr = [
             x.item() for x in flow.linspace(0, drop_path_rate, depth)
         ]  # stochastic depth decay rule
         self.blocks = nn.Sequential(
             *[
-                Block(
-                    dim=embed_dim,
-                    num_heads=num_heads,
-                    mlp_ratio=mlp_ratio,
-                    qkv_bias=qkv_bias,
-                    drop=drop_rate,
-                    attn_drop=attn_drop_rate,
-                    drop_path=dpr[i],
-                    norm_layer=norm_layer,
-                    act_layer=act_layer,
-                )
-                for i in range(depth)
+                TransformerLayer(
+                    hidden_size=embed_dim,
+                    ffn_hidden_size=ffn_size,
+                    num_attention_heads=num_heads,
+                    attention_dropout_prob=attn_drop_rate,
+                    output_dropout_prob=drop_rate,
+                    drop_path_prob=dpr[i],
+                    layer_idx=i
+                ) for i in range(depth)
             ]
         )
-        self.norm = norm_layer(embed_dim)
-
-        # Representation layer
-        if representation_size and not distilled:
-            self.num_features = representation_size
-            self.pre_logits = nn.Sequential(
-                OrderedDict(
-                    [
-                        ("fc", nn.Linear(embed_dim, representation_size)),
-                        ("act", nn.Tanh()),
-                    ]
-                )
-            )
-        else:
-            self.pre_logits = nn.Identity()
-
-        # Classifier head(s)
-        self.head = nn.Linear(self.num_features, num_classes) if num_classes > 0 else nn.Identity()
-        self.head_dist = None
-        if distilled:
-            self.head_dist = (
-                nn.Linear(self.embed_dim, self.num_classes) if num_classes > 0 else nn.Identity()
-            )
+        self.norm = LayerNorm(embed_dim)
+        self.head = Linear(embed_dim, num_classes)
 
         # Loss func
         self.loss_func = nn.CrossEntropyLoss() if loss_func is None else loss_func
 
-        self.init_weights(weight_init)
+        # weight init
+        trunc_normal_(self.pos_embed, std=.02)
+        trunc_normal_(self.cls_token, std=.02)
+        self.apply(self._init_weights)
+
+    def _init_weights(self, m):
+        if isinstance(m, Linear):
+            trunc_normal_(m.weight, std=.02)
+            if isinstance(m, Linear) and m.bias is not None:
+                nn.init.constant_(m.bias, 0)
+        elif isinstance(m, LayerNorm):
+            nn.init.constant_(m.bias, 0)
+            nn.init.constant_(m.weight, 1.0)
 
     @classmethod
-    def from_config(cls, cfg):
+    def from_config(self, cfg):
         return {
             "img_size": cfg.img_size,
             "patch_size": cfg.patch_size,
             "in_chans": cfg.in_chans,
-            "num_classes": cfg.num_classes,
             "embed_dim": cfg.embed_dim,
             "depth": cfg.depth,
             "num_heads": cfg.num_heads,
             "mlp_ratio": cfg.mlp_ratio,
-            "qkv_bias": cfg.qkv_bias,
             "drop_rate": cfg.drop_rate,
             "attn_drop_rate": cfg.attn_drop_rate,
             "drop_path_rate": cfg.drop_path_rate,
+            "num_classes": cfg.num_classes,
+            "loss_func": cfg.loss_func,
         }
 
-    def init_weights(self, mode=""):
-        assert mode in ("jax", "jax_nlhb", "nlhb", "")
-        head_bias = -math.log(self.num_classes) if "nlhb" in mode else 0.0
-        trunc_normal_(self.pos_embed, std=0.02)
-        if self.dist_token is not None:
-            trunc_normal_(self.dist_token, std=0.02)
-        if mode.startswith("jax"):
-            # leave cls token as zeros to match jax impl
-            named_apply(partial(_init_vit_weights, head_bias=head_bias, jax_impl=True), self)
-        else:
-            trunc_normal_(self.cls_token, std=0.02)
-            self.apply(_init_vit_weights)
-
-    def _init_weights(self, m):
-        # this fn left here for compat with downstream users
-        _init_vit_weights(m)
 
     def forward_features(self, x):
-        # position embedding
+        # patch embedding
         x = self.patch_embed(x)
 
         cls_token = self.cls_token.expand(
             x.shape[0], -1, -1
         )  # stole cls_tokens impl from Phil Wang, thanks
         cls_token = cls_token.to_global(sbp=flow.sbp.split(0), placement=cls_token.placement)
-        if self.dist_token is None:
-            x = flow.cat((cls_token, x), dim=1)
-        else:
-            x = flow.cat(
-                (cls_token, self.dist_token.expand(x.shape[0], -1, -1), x).to_global(
-                    sbp=flow.sbp.split(0), placement=self.dist_token.placement
-                ),
-                dim=1,
-            )
+        x = flow.cat((cls_token, x), dim=1)
+
+        # position embedding
         pos_embed = self.pos_embed.expand(x.shape[0], -1, -1)
         pos_embed = pos_embed.to_global(sbp=flow.sbp.split(0), placement=pos_embed.placement)
         x = self.pos_drop(x + self.pos_embed)
-        # transformer encoder
+
+        # transformer block
         x = self.blocks(x)
         x = self.norm(x)
 
-        if self.dist_token is None:
-            return self.pre_logits(x[:, 0])
-        else:
-            return x[:, 0], x[:, 1]
-
+        return x[:, 0]
+    
     def forward(self, images, labels=None):
         x = self.forward_features(images)
-        # classification head
-        if self.head_dist is not None:
-            x, x_dist = self.head(x[0]), self.head_dist(x[1])  # x must be a tuple
-            if self.training:
-                # during inference, return the average of both classifier predictions
-                return x, x_dist
-            else:
-                return (x + x_dist) / 2
-        else:
-            x = self.head(x)
+        x = self.head(x)
 
         if labels is not None and self.training:
             losses = self.loss_func(x, labels)
             return {"losses": losses}
         else:
             return {"prediction_scores": x}
-
-
-def _init_vit_weights(
-    module: nn.Module, name: str = "", head_bias: float = 0.0, jax_impl: bool = False
-):
-    """ViT weight initialization
-    * When called without n, head_bias, jax_impl args it will behave exactly the same
-      as my original init for compatibility with prev hparam / downstream use cases (ie DeiT).
-    * When called w/ valid n (module name) and jax_impl=True, will (hopefully) match JAX impl
-    """
-    if isinstance(module, nn.Linear):
-        if name.startswith("head"):
-            nn.init.zeros_(module.weight)
-            nn.init.constant_(module.bias, head_bias)
-        elif name.startswith("pre_logits"):
-            lecun_normal_(module.weight)
-            nn.init.zeros_(module.bias)
-        else:
-            if jax_impl:
-                nn.init.xavier_uniform_(module.weight)
-                if module.bias is not None:
-                    if "mlp" in name:
-                        nn.init.normal_(module.bias, std=1e-6)
-                    else:
-                        nn.init.zeros_(module.bias)
-            else:
-                trunc_normal_(module.weight, std=0.02)
-                if module.bias is not None:
-                    nn.init.zeros_(module.bias)
-    elif jax_impl and isinstance(module, nn.Conv2d):
-        lecun_normal_(module.weight)
-        if module.bias is not None:
-            nn.init.zeros_(module.bias)
-    elif isinstance(module, (nn.LayerNorm, nn.GroupNorm, nn.BatchNorm2d)):
-        nn.init.zeros_(module.bias)
-        nn.init.ones_(module.weight)
